ajout de fonction payer

+++ app/src/main/java/com/carrental/database/DatabaseHelper.java
@@ -8,13 +8,14 @@ import android.content.ContentValues;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     private static final String DATABASE_NAME = "car_rental.db";
-    private static final int DATABASE_VERSION = 3;
+    private static final int DATABASE_VERSION = 4;
 
     // Table names
     public static final String TABLE_USER = "Utilisateur";
     public static final String TABLE_CAR = "Voiture";
     public static final String TABLE_CLIENT = "Client";
     public static final String TABLE_RESERVATION = "Reservation";
+    public static final String TABLE_PAYMENT = "Payment";
 
     // User table columns
     public static final String COL_USER_ID = "id";
@@ -37,6 +38,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     public static final String COL_CLIENT_EMAIL = "email";
     public static final String COL_CLIENT_TELEPHONE = "telephone";
     public static final String COL_CLIENT_USER_ID = "user_id";
+    public static final String COL_CLIENT_SOLDE = "solde";
 
     // Reservation table columns
     public static final String COL_RES_ID = "id";
@@ -46,6 +48,13 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     public static final String COL_RES_DATE_FIN = "date_fin";
     public static final String COL_RES_PRIX_TOTAL = "prix_total";
     public static final String COL_RES_STATUT = "statut";
+
+    // Payment table columns
+    public static final String COL_PAY_ID = "id";
+    public static final String COL_PAY_RES_ID = "reservation_id";
+    public static final String COL_PAY_CLIENT_ID = "client_id";
+    public static final String COL_PAY_AMOUNT = "amount";
+    public static final String COL_PAY_DATE = "payment_date";
 
     private static final String CREATE_USER_TABLE =
             "CREATE TABLE " + TABLE_USER + " (" +
@@ -71,6 +80,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                     COL_CLIENT_EMAIL + " TEXT UNIQUE NOT NULL, " +
                     COL_CLIENT_TELEPHONE + " TEXT NOT NULL, " +
                     COL_CLIENT_USER_ID + " INTEGER UNIQUE, " +
+                    COL_CLIENT_SOLDE + " REAL DEFAULT 0, " +
                     "FOREIGN KEY(" + COL_CLIENT_USER_ID + ") REFERENCES " + TABLE_USER + "(" + COL_USER_ID + "))";
 
     private static final String CREATE_RESERVATION_TABLE =
@@ -85,6 +95,16 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                     "FOREIGN KEY(" + COL_RES_CLIENT_ID + ") REFERENCES " + TABLE_CLIENT + "(" + COL_CLIENT_ID + "), " +
                     "FOREIGN KEY(" + COL_RES_CAR_ID + ") REFERENCES " + TABLE_CAR + "(" + COL_CAR_ID + "))";
 
+    private static final String CREATE_PAYMENT_TABLE =
+            "CREATE TABLE " + TABLE_PAYMENT + " (" +
+                    COL_PAY_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
+                    COL_PAY_RES_ID + " INTEGER NOT NULL, " +
+                    COL_PAY_CLIENT_ID + " INTEGER NOT NULL, " +
+                    COL_PAY_AMOUNT + " REAL NOT NULL, " +
+                    COL_PAY_DATE + " TEXT NOT NULL, " +
+                    "FOREIGN KEY(" + COL_PAY_RES_ID + ") REFERENCES " + TABLE_RESERVATION + "(" + COL_RES_ID + "), " +
+                    "FOREIGN KEY(" + COL_PAY_CLIENT_ID + ") REFERENCES " + TABLE_CLIENT + "(" + COL_CLIENT_ID + "))";
+
     public DatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
     }
@@ -95,6 +115,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         db.execSQL(CREATE_CAR_TABLE);
         db.execSQL(CREATE_CLIENT_TABLE);
         db.execSQL(CREATE_RESERVATION_TABLE);
+        db.execSQL(CREATE_PAYMENT_TABLE);
 
         // Insert admin user
         ContentValues values = new ContentValues();
@@ -124,6 +145,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         values.put(COL_CLIENT_EMAIL, "sophie.martin@email.com");
         values.put(COL_CLIENT_TELEPHONE, "0612345678");
         values.put(COL_CLIENT_USER_ID, clientUserId);
+        values.put(COL_CLIENT_SOLDE, 1000.0); // Sample balance
         long clientId = db.insert(TABLE_CLIENT, null, values);
 
         // Insert sample cars
@@ -143,14 +165,6 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         values.put(COL_CAR_DISPONIBLE, 1);
         long carId2 = db.insert(TABLE_CAR, null, values);
 
-        values = new ContentValues();
-        values.put(COL_CAR_MARQUE, "Citroën");
-        values.put(COL_CAR_MODELE, "C3");
-        values.put(COL_CAR_ANNEE, 2021);
-        values.put(COL_CAR_PRIX_JOUR, 45.0);
-        values.put(COL_CAR_DISPONIBLE, 1);
-        db.insert(TABLE_CAR, null, values);
-
         // Insert sample reservation
         values = new ContentValues();
         values.put(COL_RES_CLIENT_ID, clientId);
@@ -164,10 +178,16 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_RESERVATION);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_CLIENT);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_CAR);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_USER);
-        onCreate(db);
+        if (oldVersion < 4) {
+            db.execSQL("ALTER TABLE " + TABLE_CLIENT + " ADD COLUMN " + COL_CLIENT_SOLDE + " REAL DEFAULT 0");
+            db.execSQL(CREATE_PAYMENT_TABLE);
+        } else {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_PAYMENT);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_RESERVATION);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_CLIENT);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_CAR);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_USER);
+            onCreate(db);
+        }
     }
 }
